107_Document_Loaders/requirements.txt
# LangChain Core
langchain>=0.1.14
langchain-core>=0.1.37

# OpenAI Integration
openai>=1.27.0

#Pydantic for Data Validation
pydantic>=2.6.4

#Environment Variables Management
python-dotenv

#Web Frameworks
streamlit

langchain_community
# Document Loaders
pypdf